+++ src/main/java/ApplicationBusiness/DataAccessInterface.java
@@ -1,6 +1,8 @@
+package ApplicationBusiness;
+
 import java.util.ArrayList;
 
-interface DataAccessInterface {
+public interface DataAccessInterface {
     ArrayList<String> read();
     void write(ArrayList<String> data);
 }
