+++ src/main/java/ApplicationBusiness/EasyUseCase.java
@@ -1,0 +1,4 @@
+package ApplicationBusiness;
+
+public class EasyUseCase {
+}
