+++ src/main/java/ApplicationBusiness/GameOutputBoundary.java
@@ -1,3 +1,5 @@
+package ApplicationBusiness;
+
 public interface GameOutputBoundary {
     void updateGame(String position, int time, int score);
 }
