+++ src/main/java/ApplicationBusiness/HardUseCase.java
@@ -1,0 +1,4 @@
+package ApplicationBusiness;
+
+public class HardUseCase {
+}
