+++ src/main/java/ApplicationBusiness/InterfaceLeaderboardPresenter.java
@@ -1,0 +1,7 @@
+package ApplicationBusiness;
+
+import java.util.ArrayList;
+
+public interface InterfaceLeaderboardPresenter {
+    void setData(ArrayList<String> data);
+}
