+++ src/main/java/ApplicationBusiness/MediumUseCase.java
@@ -1,2 +1,4 @@
+package ApplicationBusiness;
+
 public class MediumUseCase {
 }
