+++ src/main/java/FrameworksAndDrivers/EasyUI.java
@@ -1,6 +1,10 @@
+package FrameworksAndDrivers;
+
+import InterfaceAdapters.NavigatorController;
+import InterfaceAdapters.ViewModel;
+
 import javax.swing.*;
 import java.awt.*;
-import java.awt.Window;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 import java.util.Objects;
