+++ src/main/java/FrameworksAndDrivers/Leaderboard_Frame.java
@@ -1,3 +1,7 @@
+package FrameworksAndDrivers;
+
+import InterfaceAdapters.ViewModel;
+
 import javax.swing.*;
 
 public class Leaderboard_Frame extends JFrame {
