--- src/main/java/HardUseCase.java
@@ -1,2 +1,0 @@
-public class HardUseCase {
-}
