--- src/main/java/InterfaceLeaderboardPresenter.java
@@ -1,5 +1,0 @@
-import java.util.ArrayList;
-
-interface InterfaceLeaderboardPresenter {
-    void setData(ArrayList<String> data);
-}
